Split save and cry files + some doc
