+++ src/save.c
@@ -4,98 +4,31 @@
 #include "main.h"
 #include "constants/species_rs.h"
 
-extern const u16 gSpeciesRSToCryId[];
-
-static void PlayCryInternal(u16, s8, s8, u8, int);
-static u16 LoadSaveDataFromSram(void);
-
-int SpeciesRSToCryId(u16 speciesRS)
-{
-    if (speciesRS <= SPECIES_RS_CELEBI - 1)
-        return speciesRS;
-    if (speciesRS < SPECIES_RS_TREECKO - 1)
-        return SPECIES_RS_UNOWN - 1;
-    return gSpeciesRSToCryId[speciesRS - (SPECIES_RS_TREECKO - 1)];
-}
-
-void PlayCry_Normal(u16 speciesRS, s8 pan)
-{
-    m4aMPlayVolumeControl(&gMPlayInfo_BGM, TRACKS_ALL, 153);
-    PlayCryInternal(speciesRS, pan, 125, 10, 0);
-}
-
-void PlayCry_NormalNoDucking(u16 speciesRS, s8 pan, s8 volume, u8 priority)
-{
-    PlayCryInternal(speciesRS, pan, volume, priority, 0);
-}
-
-// Probably was ported from RS minus the cry mode section.
-static void PlayCryInternal(u16 speciesRS, s8 pan, s8 volume, u8 priority, int unused)
-{
-    u32 release;
-    u32 length;
-    u32 pitch;
-    u32 var;
-    u32 index;
-    u8 table;
-
-    speciesRS--;
-
-    length = 140;
-    release = 0;
-    pitch = 15360;
-
-    SetPokemonCryVolume(volume);
-    SetPokemonCryPanpot(pan);
-    SetPokemonCryPitch(pitch);
-    SetPokemonCryLength(length);
-    SetPokemonCryProgress(0);
-    SetPokemonCryRelease(release);
-    SetPokemonCryChorus(0);
-    SetPokemonCryPriority(priority);
-    var = SpeciesRSToCryId(speciesRS);
-    index = var & 0x7F;
-    table = var >> 7;
-    switch (table)
-    {
-    case 0:
-        SetPokemonCryTone(&gUnknown_08532D6C[index]);
-        break;
-    case 1:
-        SetPokemonCryTone(&gUnknown_08533360[index]);
-        break;
-    case 2:
-        SetPokemonCryTone(&gUnknown_08533960[index]);
-        break;
-    case 3:
-        SetPokemonCryTone(&gUnknown_08533F60[index]);
-        break;
-    }
-}
+static bool16 LoadSaveDataFromSram(void);
 
 void SaveFile_LoadGameData(void)
 {
     SetSramFastFunc();
     gMain.unkC = 0;
-    if (LoadSaveDataFromSram() == 0)
+    if (LoadSaveDataFromSram() == FALSE)
     {
-        sub_52C64();
+        ResetSaveFile();
         SaveFile_WriteToSram();
-        if (LoadSaveDataFromSram() == 0)
+        if (LoadSaveDataFromSram() == FALSE)
         {
             gMain.unkC = 1;
-            sub_52C64();
+            ResetSaveFile();
         }
     }
     else
     {
-        sub_525CC(gMain_saveData.buttonConfigType);
+        SetButtonConfigInputs(gMain_saveData.buttonConfigType);
     }
 }
 
 extern u8 gSaveFileSignature[];
 
-static u16 LoadSaveDataFromSram(void)
+static bool16 LoadSaveDataFromSram(void)
 {
     u16 isOk = FALSE;
     u16 fileNum;
@@ -195,7 +128,7 @@ void SaveFile_ReadSavedGamePresent(void)
     ReadSramFast((void *)(SRAM + 0x544), (u8 *)&gMain.hasSavedGame, sizeof(gMain.hasSavedGame));
 }
 
-void sub_52C64(void)
+void ResetSaveFile(void)
 {
     s16 i;
 
@@ -205,7 +138,7 @@ void sub_52C64(void)
     gMain_saveData.unk2E4 = 0;
     gMain_saveData.rumbleEnabled = FALSE;
     gMain_saveData.ballSpeed = 0;
-    sub_525CC(BUTTON_CONFIG_RESET);
+    SetButtonConfigInputs(BUTTON_CONFIG_RESET);
     SetDefaultHighScores();
     ResetPokedex();
     gMain_saveData.buttonConfigType = BUTTON_CONFIG_TYPE_A;
